refactor: move app configs to separate files
