+++ src/config/swagger.config.ts
@@ -1,0 +1,13 @@
+import { INestApplication } from '@nestjs/common';
+import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
+
+export const createSwaggerSchema = (app: INestApplication) => {
+  const config = new DocumentBuilder()
+    .setTitle('Node.js Home Library Service')
+    .setVersion('1.0')
+    .build();
+
+  const document = SwaggerModule.createDocument(app, config);
+
+  SwaggerModule.setup('doc', app, document);
+};
